+++ COT/helpers/mkisofs.py
@@ -3,7 +3,7 @@
 # mkisofs.py - Helper for 'mkisofs' and 'genisoimage'
 #
 # February 2015, Glenn F. Matthews
-# Copyright (c) 2013-2016 the COT project developers.
+# Copyright (c) 2013-2017 the COT project developers.
 # See the COPYRIGHT.txt file at the top-level directory of this distribution
 # and at https://github.com/glennmatthews/cot/blob/master/COPYRIGHT.txt.
 #
@@ -39,6 +39,7 @@ class MkISOFS(Helper):
     """
 
     _provider_package = {
+        'brew': 'cdrtools',
         'port': 'cdrtools',
     }